Move guicolors to gui package (more universal colors rather than just scene)
